Release 1.0.17
longport@1.0.17
longport-c@1.0.17
longport-c-macros@1.0.17
longport-candlesticks@1.0.17
longport-httpcli@1.0.17
longport-java@1.0.17
longport-java-macros@1.0.17
longport-nodejs@1.0.17
longport-nodejs-macros@1.0.17
longport-proto@1.0.17
longport-python@1.0.17
longport-python-macros@1.0.17
longport-wscli@1.0.17

+++ cpp/include/types.hpp
@@ -1535,7 +1535,7 @@ struct StockPosition
   /// Market
   Market market;
   /// Initial position before market opening
-  std::optional<Decimal> init_quantity;
+  std::optional<int64_t> init_quantity;
 };
 
 /// Stock position channel
